BOJAN-006
Fixing ratings

+++ green-health/src/main/java/com/green/health/parents/RatingsParent.java
@@ -52,7 +52,7 @@ public interface RatingsParent extends PojoParent {
 	
 	default public double calculateRating() {
 		long total = getRatingOnes() + getRatingTwos() + getRatingThrees() + getRatingFours() + getRatingFives();
-		return (getRatingOnes() + 2*getRatingTwos() + 3*getRatingThrees() + 4*getRatingFours() + 5*getRatingFives())/
+		return (double)(getRatingOnes() + 2*getRatingTwos() + 3*getRatingThrees() + 4*getRatingFours() + 5*getRatingFives())/
 				(total!=0 ? total : 1);
 	}
 }
